+++ psycop/common/global_utils/mlflow/mlflow_data_extraction.py
@@ -1,9 +1,14 @@
-from collections.abc import Iterable
+from collections.abc import Sequence
 from dataclasses import dataclass
 from pathlib import Path
 
+import mlflow
 import polars as pl
+from confection import Config
 from mlflow.entities import Run
+from mlflow.entities.run_data import RunData
+from mlflow.entities.run_info import RunInfo
+from mlflow.entities.run_inputs import RunInputs
 from mlflow.tracking import MlflowClient
 
 from psycop.common.types.validated_frame import ValidatedFrame
@@ -20,86 +25,72 @@ class MlflowAllMetricsFrame(ValidatedFrame[pl.DataFrame]):
     allow_extra_columns = False
 
 
-@dataclass(frozen=True)
-class MlflowTimeVaryingMetricsFrame(ValidatedFrame[pl.DataFrame]):
-    frame: pl.DataFrame
+class PsycopMlflowRun(Run):
+    def __init__(
+        self,
+        run_info: RunInfo,
+        run_data: RunData,
+        run_inputs: RunInputs | None,
+        client: MlflowClient,
+    ) -> None:
+        super().__init__(run_info, run_data, run_inputs)
+        self._client = client
 
-    run_name_col_name: str = "run_name"
-    metric_col_name: str = "metric"
-    value_col_name: str = "value"
-    unix_timestamp_col_name: str = "timestamp"
-    step_col_name: str = "step"
+    @classmethod
+    def from_mlflow_run(
+        cls: type["PsycopMlflowRun"], run: Run, client: MlflowClient
+    ) -> "PsycopMlflowRun":
+        return cls(run_info=run._info, run_data=run._data, run_inputs=run._inputs, client=client)
 
-    allow_extra_columns = False
+    def get_config(self) -> Config:
+        cfg_path = self.download_artifact(artifact_name="config.cfg", save_location=None)
+        return Config().from_disk(cfg_path)
+
+    def download_artifact(self, artifact_name: str, save_location: str | None = None) -> Path:
+        """Download an artifact from a run. Returns the path to the downloaded artifact.
+        If save_location is None, will save to temporary directory"""
+        return Path(
+            self._client.download_artifacts(
+                run_id=self.info.run_id, path=artifact_name, dst_path=save_location
+            )
+        )
+
+    def get_all_metrics(self) -> pl.DataFrame:
+        return pl.DataFrame(self.data.metrics).with_columns(
+            pl.lit(self.info.run_name).alias("run_name")
+        )
 
 
-class MlflowMetricExtractor:
+class MlflowClientWrapper:
     def __init__(self) -> None:
-        self.client = MlflowClient(tracking_uri="http://exrhel0371.it.rm.dk:5050")
+        tracking_uri = "http://exrhel0371.it.rm.dk:5050"
+
+        self.client = MlflowClient(tracking_uri=tracking_uri)
+        mlflow.set_tracking_uri(tracking_uri)
 
     def get_all_metrics_for_experiment(self, experiment_name: str) -> MlflowAllMetricsFrame:
         """Get the final value of all logged metrics for each run in an experiment.
         Returns a long df with cols 'run_name', 'metric', and 'value'."""
         runs = self._get_mlflow_runs_by_experiment(experiment_name=experiment_name)
-        metrics_df = pl.concat([self._get_all_metrics_for_run(run) for run in runs]).melt(
+        metrics_df = pl.concat([run.get_all_metrics() for run in runs]).melt(
             id_vars="run_name", variable_name="metric"
         )
 
         return MlflowAllMetricsFrame(frame=metrics_df, allow_extra_columns=False)
 
-    def get_metrics_for_experiment(
-        self, experiment_name: str, metrics: Iterable[str]
-    ) -> MlflowTimeVaryingMetricsFrame:
-        """Get all values of specified metrics for each run in an experiment. This
-        includes values that get updated such as loss. Returns a long df with
-        cols 'run_name', 'metric', 'value', 'timestamp', 'step'."""
-        runs = self._get_mlflow_runs_by_experiment(experiment_name=experiment_name)
-        metrics_df = pl.concat(
-            [self._get_metrics_for_run(run=run, metrics=metrics) for run in runs]
-        ).rename({"key": "metric"})
-
-        return MlflowTimeVaryingMetricsFrame(frame=metrics_df, allow_extra_columns=False)
-
-    def get_best_run_from_experiments(self, experiment_names: Iterable[str], metric: str) -> Run:
+    def get_best_run_from_experiment(
+        self, experiment_name: str, metric: str, larger_is_better: bool = True
+    ) -> PsycopMlflowRun:
         """Get the best run from one or more experiments based on some metric,
         e.g. 'all_oof_BinaryAUROC'"""
-        experiment_ids = [
-            self._get_mlflow_experiment_id_from_experiment_name(experiment_name)
-            for experiment_name in experiment_names
-        ]
+        order = "DESC" if larger_is_better else "ASC"
+
+        experiment_id = self._get_mlflow_experiment_id_from_experiment_name(experiment_name)
 
         best_run = self.client.search_runs(
-            experiment_ids=experiment_ids, max_results=1, order_by=[f"metrics.{metric} DESC"]
+            experiment_ids=[experiment_id], max_results=1, order_by=[f"metrics.{metric} {order}"]
         )[0]
-        return best_run
-
-    def download_config_from_best_run_from_experiments(
-        self, experiment_names: Iterable[str], metric: str
-    ) -> Path:
-        """Download the config from the best from a list of experiments. Returns the path to the config"""
-        best_run = self.get_best_run_from_experiments(
-            experiment_names=experiment_names, metric=metric
-        )
-        return Path(self.client.download_artifacts(run_id=best_run.info.run_id, path="config.cfg"))
-
-    def _get_metrics_for_run(self, run: Run, metrics: Iterable[str]) -> pl.DataFrame:
-        metrics_df = [
-            self._get_metric_for_run(run_id=run.info.run_id, metric=metric) for metric in metrics
-        ]
-        return pl.concat(metrics_df).with_columns(pl.lit(run.info.run_name).alias("run_name"))
-
-    def _get_metric_for_run(self, run_id: str, metric: str) -> pl.DataFrame:
-        return pl.DataFrame(
-            [
-                dict(metric_object)
-                for metric_object in self.client.get_metric_history(run_id=run_id, key=metric)
-            ]
-        )
-
-    def _get_all_metrics_for_run(self, run: Run) -> pl.DataFrame:
-        return pl.DataFrame(run.data.metrics).with_columns(
-            pl.lit(run.info.run_name).alias("run_name")
-        )
+        return PsycopMlflowRun.from_mlflow_run(run=best_run, client=self.client)
 
     def _get_mlflow_experiment_id_from_experiment_name(self, experiment_name: str) -> str:
         experiment = self.client.get_experiment_by_name(name=experiment_name)
@@ -107,18 +98,21 @@ class MlflowMetricExtractor:
             raise ValueError(f"{experiment_name} does not exist on MlFlow.")
         return experiment.experiment_id
 
-    def _get_mlflow_runs_by_experiment(self, experiment_name: str) -> Iterable[Run]:
+    def _get_mlflow_runs_by_experiment(self, experiment_name: str) -> Sequence[PsycopMlflowRun]:
         experiment_id = self._get_mlflow_experiment_id_from_experiment_name(
             experiment_name=experiment_name
         )
-        return self.client.search_runs(experiment_ids=[experiment_id])
+        runs = self.client.search_runs(experiment_ids=[experiment_id])
+        return [PsycopMlflowRun.from_mlflow_run(run=run, client=self.client) for run in runs]
 
 
 if __name__ == "__main__":
-    df = MlflowMetricExtractor().get_all_metrics_for_experiment("text_exp")
-    df = MlflowMetricExtractor().get_metrics_for_experiment(
-        "text_exp", ["all_oof_BinaryAUROC", "within_fold_0_BinaryAUROC"]
-    )
-    best_run = MlflowMetricExtractor().get_best_run_from_experiments(
-        experiment_names=["text_exp"], metric="all_oof_BinaryAUROC"
+    df = MlflowClientWrapper().get_all_metrics_for_experiment("text_exp")
+
+    best_config = (
+        MlflowClientWrapper()
+        .get_best_run_from_experiment(
+            experiment_name="scz-bp_3_year_lookahead", metric="all_oof_BinaryAUROC"
+        )
+        .get_config()
     )